+++ src/main/kotlin/cloud/glitchdev/rfu/constants/MessageTypes.kt
@@ -3,6 +3,7 @@ package cloud.glitchdev.rfu.constants
 enum class MessageTypes(val displayName : String) {
     HYPE("Hyperion"),
     AUTOPET("Autopet"),
+    LOOTSHARE("Lootshare"),
     CATCH("Catch"),
     COMBO("Combo"),
     BLOCKS("Blocks in the way");