+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added Lootshare detection messages.
 - Added option to hide "There are blocks in the way!" message
 - Added option to hide Combo messages (Won't hide the MF ones)
+- Added option to hide Lootshare Messages
 - Added Zoom on etherwarp (I just really missed that)
 - Added option to make the health bar blue when the mob is shurikened (On by default)
 - Added a Dye Display HUD element